Changed screen naming due to server refactor
* Renamed [Chat] to [Message].
* Renamed [Message] to [Channel].
