+++ assets/langs/en.json
@@ -15,7 +15,7 @@
   "REGISTER": "Register",
   "SEND_LINK": "Send Link",
   "FRIEND": "Friend",
-  "MESSAGE": "Message",
+  "CHANNEL": "Channel",
   "MY_PROFILE": "My profile",
   "UPDATE": "Update",
   "NO_CONTENT": "No content",
@@ -23,6 +23,7 @@
   "DELETE_FRIEND": "Delete",
   "SEARCH_USER": "Search User",
   "CHAT": "Chat",
+  "MESSAGE": "Message",
   "SEND": "Send",
   "WRITE_MESSAGE": "Write message...",
   "ERROR": "Error",
