+++ assets/langs/ko.json
@@ -15,7 +15,7 @@
   "REGISTER": "가입",
   "SEND_LINK": "링크 보내기",
   "FRIEND": "친구",
-  "MESSAGE": "메세지",
+  "CHANNEL": "채널",
   "MY_PROFILE": "나의 프로필",
   "UPDATE": "수정",
   "NO_CONTENT": "컨텐츠가 없습니다",
@@ -23,6 +23,7 @@
   "DELETE_FRIEND": "친구삭제",
   "SEARCH_USER": "사용자 검색",
   "CHAT": "채팅",
+  "MESSAGE": "메시지",
   "SEND": "보내기",
   "WRITE_MESSAGE": "메세지를 입력해주세요…",
   "ERROR": "에러",
