+++ src/components/screen/__tests__/Message.test.tsx
@@ -1,59 +1,114 @@
 import 'react-native';
 
-import * as React from 'react';
+import * as ProfileContext from '../../../providers/ProfileModalProvider';
 
-import { RenderResult, cleanup, render } from '@testing-library/react-native';
+import React, { ReactElement } from 'react';
+import {
+  RenderResult,
+  act,
+  cleanup,
+  fireEvent,
+  render,
+} from '@testing-library/react-native';
 import { createTestElement, createTestProps } from '../../../../test/testUtils';
 
 import Message from '../Message';
+// Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 let props: any;
-let component: React.ReactElement;
-let testingLib: RenderResult;
+let component: ReactElement;
 
-jest.mock('@react-navigation/native', () => {
-  return {
-    useNavigation: (): object => {
-      return {
-        navigation: {
-          navigate: jest.fn(),
-        },
-      };
-    },
-  };
-});
+jest.mock('expo-permissions', () => ({
+  askAsync: (): string => 'granted',
+}));
 
-describe('[Message] screen', () => {
+jest.mock('expo-image-picker', () => ({
+  launchCameraAsync: (): string => 'photo info',
+  launchImageLibraryAsync: (): string => 'photo info',
+}));
+
+describe('[Message] rendering test', () => {
   beforeEach(() => {
     props = createTestProps();
     component = createTestElement(<Message {...props} />);
+  });
+
+  it('renders as expected', () => {
+    const json = renderer.create(component).toJSON();
+    expect(json).toMatchSnapshot();
+  });
+});
+
+describe('[Message] interaction', () => {
+  let testingLib: RenderResult;
+
+  beforeAll(() => {
     testingLib = render(component);
   });
 
-  it('renders without crashing', () => {
-    const rendered = renderer.create(component).toJSON();
-    expect(rendered).toMatchSnapshot();
-    expect(rendered).toBeTruthy();
-  });
-
-  describe('interactions', () => {
-    beforeEach(() => {
-      testingLib = render(component);
-    });
-
-    it('should simulate onPress', () => {
-      // const btn = testingLib.queryByTestId('btn');
-      // act(() => {
-      //   fireEvent.press(btn);
-      //   fireEvent.press(btn);
-      // });
-      // expect(cnt).toBe(3);
-    });
-  });
-
-  afterEach(() => {
+  afterAll(() => {
     cleanup();
+  });
+
+  it('should [sendMessage] when pressing button', () => {
+    let MessageBtn = testingLib.getByTestId('btn-message');
+    MessageBtn = testingLib.getByTestId('btn-message');
+    fireEvent.press(MessageBtn);
+  });
+
+  describe('dispatch showModal', () => {
+    it('should dispatch [show-modal] when peerImage is pressed', () => {
+      jest
+        .spyOn(ProfileContext, 'useProfileContext')
+        .mockImplementation(() => ({
+          showModal: jest.fn(),
+          state: null,
+        }));
+      const MessageListItem = testingLib.queryByTestId('message-list-item0');
+      act(() => {
+        fireEvent.press(MessageListItem);
+      });
+    });
+
+    it('should call [show-modal] when modal is available', () => {
+      const mockedData = {
+        showModal: jest.fn(),
+        state: {
+          user: null,
+          deleteMode: true,
+          modal: jest.mock,
+        },
+      };
+      jest
+        .spyOn(ProfileContext, 'useProfileContext')
+        .mockImplementation(() => mockedData);
+      const MessageListItem = testingLib.queryByTestId('message-list-item0');
+      testingLib.rerender(component);
+      act(() => {
+        fireEvent.press(MessageListItem);
+      });
+      expect(mockedData.showModal).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  it('should open image library when pressing photo icon button', () => {
+    const touchMenu = testingLib.getByTestId('touch_menu');
+    act(() => {
+      fireEvent.press(touchMenu);
+    });
+    const photoBtn = testingLib.getByTestId('icon_photo');
+
+    act(() => {
+      fireEvent.press(photoBtn);
+    });
+  });
+
+  it('should open camera when pressing camera icon button', () => {
+    const cameraBtn = testingLib.getByTestId('icon_camera');
+    act(() => {
+      fireEvent.press(cameraBtn);
+    });
   });
 });
