+++ src/components/screen/__tests__/Setting.test.tsx
@@ -23,7 +23,7 @@ function getEmptyAuthUserWithSignInType(signInType: SocialType): AuthUser {
     photoURL: '',
     statusMsg: '',
     friends: [],
-    chatrooms: [],
+    channels: [],
     social: signInType,
   };
 }
