+++ src/theme.ts
@@ -45,7 +45,7 @@ export const light = {
   primaryLight: colors.paleBlue_10,
   headerFont: colors.light,
   searchBackground: colors.gray_05,
-  chatBackground: colors.light,
+  messageBackground: colors.light,
   status: colors.green_10,
   placeholder: colors.gray_60,
   focused: colors.blue_70,
@@ -91,7 +91,7 @@ export const dark = {
   primaryLight: colors.dark,
   headerFont: colors.light,
   searchBackground: colors.gray_80,
-  chatBackground: colors.dark,
+  messageBackground: colors.dark,
   status: colors.green_30,
   placeholder: colors.gray_50,
   focused: '#ffffff',
